backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
inflection==0.5.1
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
